Address canvas panel event review
